requirements.txt
Click==7.0
python-dotenv==0.10.3
todoist-python==8.0.0
